requirements.txt
aiohttp== 3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
discord==1.7.3
discord.py==1.7.3
discord-py-slash-command==3.0.3
<<<<<<< HEAD
lxml==4.9.1
linkedin_api==1.1.0
=======
idna==3.3
multidict==6.0.2
pip==22.0.2
python-dateutil==2.8.2
requests==2.28.1
setuptools==59.6.0
typing_extensions==4.3.0
urllib3==1.26.11
yarl==1.8.1
>>>>>>> main
